services/analytics/requirements.txt
# Analytics Service Dependencies

# FastAPI and async support
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP client and async support
httpx>=0.25.0
aiofiles>=23.0.0

# Epic 1 CostTracker dependencies
# (These should match Epic 1 requirements)

# Data handling and analytics
numpy>=1.24.0
pandas>=2.0.0

# Monitoring and metrics
prometheus-client>=0.19.0
structlog>=23.0.0

# Circuit breaker pattern
pybreaker>=1.0.0

# Configuration management
PyYAML>=6.0

# Optional persistence backends
redis>=5.0.0
asyncio-redis>=0.16.0
psycopg2-binary>=2.9.0  # PostgreSQL support
sqlalchemy>=2.0.0  # ORM support

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
httpx>=0.25.0  # For testing API endpoints

# Type hints
typing-extensions>=4.5.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3